+++ packages/gamut-patterns/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.10.15](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.10.14...@codecademy/gamut-patterns@0.10.15) (2025-08-20)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+### [0.10.14](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.10.13...@codecademy/gamut-patterns@0.10.14) (2025-08-20)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+### [0.10.13](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.10.12...@codecademy/gamut-patterns@0.10.13) (2025-07-22)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+### [0.10.12](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.10.11...@codecademy/gamut-patterns@0.10.12) (2025-07-17)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
 ### [0.10.11](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.10.10...@codecademy/gamut-patterns@0.10.11) (2025-06-17)
 
 **Note:** Version bump only for package @codecademy/gamut-patterns